+++ src/sot/__about__.py
@@ -1,4 +1,4 @@
 from datetime import datetime
 
-__version__ = "6.1.0"
+__version__ = "6.2.0"
 __current_year__ = datetime.now().year